+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 1.1.5 (TBD)
+## [1.1.5](https://github.com/ngageoint/geopackage-core-java/releases/tag/1.1.5) (02-02-2016)
 
 * Feature Tile Link Extension implementation - http://ngageoint.github.io/GeoPackage/docs/extensions/feature-tile-link.html
 * GeoPackage drop table method